requirements.txt
streamlit
requests
python-dotenv
pdfplumber
